backend/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# PDF Generation
reportlab==4.0.8
matplotlib==3.8.2

# Development
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
